+++ android-client/app/src/main/java/de/njsm/stocks/android/repo/EventRepository.java
@@ -63,4 +63,13 @@ public class EventRepository {
                         eventDao.getFoodItemHistoryOfSingleFood(id))
                 , 7).build();
     }
+
+    public LiveData<PagedList<EntityEvent<?>>> getLocationEvents(int id) {
+        LOG.d("getting history");
+        return new LivePagedListBuilder<>(
+                new MergingDataSourceFactory(
+                        eventDao.getLocationHistoryOfSingleLocation(id),
+                        eventDao.getFoodItemHistoryOfSingleLocation(id))
+                , 7).build();
+    }
 }